refactor: auth credentials
